Added the course names
in each .txt file, i have added the course name as the second element in the Comma Separated Value (CSV) lines.

(course ID - Code),(Course Name)

the course name is significant to put onto here because when we want to print the list of courses or even just one course, its easy to access that string because it's at its correct place in the .txt file relative to the degree.

Merci